+++ src/components/TestRail/TestRail.js
@@ -3,12 +3,15 @@ const ApiClient = require('./ApiClient');
 
 class TestRail {
     /**
+     *
      * @param domain
      * @param username
      * @param password
+     * @param isScreenshotsEnabled
      */
-    constructor(domain, username, password) {
+    constructor(domain, username, password, isScreenshotsEnabled) {
         this.client = new ApiClient(domain, username, password);
+        this.isScreenshotsEnabled = isScreenshotsEnabled;
     }
 
     /**
@@ -125,7 +128,7 @@ class TestRail {
 
                 ColorConsole.success('  TestRail result ' + resultId + ' sent for TestCase C' + result.getCaseId());
 
-                if (result.getScreenshotPath() !== null && result.getScreenshotPath() !== '') {
+                if (this.isScreenshotsEnabled && result.getScreenshotPath() !== null && result.getScreenshotPath() !== '') {
                     ColorConsole.debug('    sending screenshot to TestRail for TestCase C' + result.getCaseId());
                     this.client.sendScreenshot(resultId, result.getScreenshotPath(), null, null);
                 }