+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added new CLI command (setup) to build your configuration in an interactive way. (see README)
 - The integration will now automatically send the screenshots to TestRail for every failed test. This is all automatically done as soon as the new variable is enabled (see README).
 - Added option to provide optional "suiteId" when creating new runs. If you have a project with suites, this parameter is required.